% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_loadings.R
\name{efa_loadings}
\alias{efa_loadings}
\title{EFA Loadings and Communalities}
\usage{
efa_loadings(fit, print = TRUE)
}
\arguments{
\item{fit}{an efa model created by psych::fa()}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of factor loadings and communalities for an EFA
}
