% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_rotmatrix.R
\name{efa_rotmatrix}
\alias{efa_rotmatrix}
\title{EFA Factor Rotation Matrix}
\usage{
efa_rotmatrix(fit, print = TRUE)
}
\arguments{
\item{fit}{an efa model created by psych::fa()}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a factor rotation matrix for an EFA
}
