% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_var.R
\name{efa_var}
\alias{efa_var}
\title{EFA Total Variance Explained}
\usage{
efa_var(fit, print = TRUE)
}
\arguments{
\item{fit}{an efa model created by psych::fa()}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of total variance explained in an EFA
}
