% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ci.R
\name{format_ci}
\alias{format_ci}
\title{Format confidence intervals for table display}
\usage{
format_ci(x, digits = digits)
}
\arguments{
\item{x}{lavaan object}

\item{digits}{How many digits to display? (default = 3)}
}
\description{
Format confidence intervals for table display
}
