% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stars.R
\name{format_stars}
\alias{format_stars}
\title{Format significant level stars for table display}
\usage{
format_stars(x)
}
\arguments{
\item{x}{lavaan object}

\item{digits}{How many digits to display? (default = 3)}
}
\description{
Format significant level stars for table display
}
