% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.descriptives.R
\name{sem.descriptives-deprecated}
\alias{sem.descriptives-deprecated}
\title{Descriptives table}
\usage{
sem.descriptives(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
\seealso{
\code{\link{semoutput-deprecated}}
}
\keyword{internal}
