% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.descriptives.R
\name{sem.descriptives}
\alias{sem.descriptives}
\title{Descriptives table}
\usage{
sem.descriptives(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
\examples{
sem.descriptives(x)
}
