% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.factorvar.R
\name{sem.factorvar}
\alias{sem.factorvar}
\title{An SEM Output Function}
\usage{
sem.factorvar(x, factors = c(), standardized = TRUE, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{factors}{list c() of factors included in the model}

\item{standardized}{logical whether to include standardized loadings (default = TRUE)}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Latent factor variances
}
\examples{
sem.factorvar(x)
}
