% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.fitmeasures.R
\name{sem.fitmeasures-deprecated}
\alias{sem.fitmeasures-deprecated}
\title{Model fit statistics}
\usage{
sem.fitmeasures(x, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Model fit measures
}
\seealso{
\code{\link{semoutput-deprecated}}
}
\keyword{internal}
