% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.residuals.R
\name{sem.residuals-deprecated}
\alias{sem.residuals-deprecated}
\title{Residual correlation matrix}
\usage{
sem.residuals(x, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of the residual correlation matrix of a lavaan cfa() or sem() model
}
\seealso{
\code{\link{semoutput-deprecated}}
}
\keyword{internal}
