% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.rsquared.R
\name{sem.rsquared}
\alias{sem.rsquared}
\title{R squared values}
\usage{
sem.rsquared(x, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of R squared values (explained variance)
}
\examples{
sem.rsquared(x)
}
