% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.sig.R
\name{sem.sig}
\alias{sem.sig}
\title{Chi-square significance test}
\usage{
sem.sig(x, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Model significance tests
}
\examples{
sem.sig(x)
}
