% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_anova.R
\name{sem_anova-deprecated}
\alias{sem_anova-deprecated}
\title{Model comparison}
\usage{
sem_anova(m1, m2, print = TRUE)
}
\arguments{
\item{m1}{A nested lavaan model}

\item{m2}{A nested lavaan model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Model significance tests
}
\seealso{
\code{\link{semoutput-deprecated}}
}
\keyword{internal}
