% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_descriptives.R
\name{sem_descriptives}
\alias{sem_descriptives}
\title{Descriptives table}
\usage{
sem_descriptives(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
