% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_factorloadings.R
\name{sem_factorloadings}
\alias{sem_factorloadings}
\title{Factor loadings}
\usage{
sem_factorloadings(
  x,
  standardized = TRUE,
  unstandardized = FALSE,
  ci_level = 0.95,
  digits = 3,
  print = TRUE
)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{standardized}{Logical, indicating whether or not to print standardized
estimates. Standardized estimates are based on "refit" of the model
on standardized data but it will not standardize categorical predictors.
Defualt is TRUE.}

\item{unstandardized}{Logical, indicating whether or not to print
unstandardized estimates. Default is TRUE.}

\item{ci_level}{What level of confidence interval to use (default = 0.95)}

\item{digits}{How many digits to display? (default = 3)}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of Factor loadings
}
