% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_factorvar.R
\name{sem_factorvar}
\alias{sem_factorvar}
\title{Latent factor variances}
\usage{
sem_factorvar(x, digits = 3, print = TRUE)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{digits}{How many digits to display? (default = 3)}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of Latent factor variances
}
