% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fitmeasures.R
\name{sem_fitmeasures}
\alias{sem_fitmeasures}
\title{Model fit statistics}
\usage{
sem_fitmeasures(x, print = TRUE)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of Model fit measures
}
