% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_intercepts.R
\name{sem_intercepts}
\alias{sem_intercepts}
\title{SEM intercepts}
\usage{
sem_intercepts(x, ci_level = 0.95, digits = 3, print = TRUE)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{ci_level}{What level of confidence interval to use (default = 0.95)}

\item{digits}{How many digits to display? (default = 3)}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of variable intercepts
}
