% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_modelcomp.R
\name{sem_modelcomp}
\alias{sem_modelcomp}
\title{Model comparison}
\usage{
sem_modelcomp(m1, m2, print = TRUE, method = "chi-square")
}
\arguments{
\item{m1}{A nested lavaan model. Null hypothesis}

\item{m2}{A nested lavaan model. Alternative hypothesis}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}

\item{method}{Model comparison method. "chi-square" or "BIC"}
}
\description{
This function will display a table of Model significance tests
}
