% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_paths.R
\name{sem_paths}
\alias{sem_paths}
\title{SEM regression paths}
\usage{
sem_paths(x, standardized = TRUE, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{standardized}{logical whether to include standardized loadings (default = TRUE)}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of SEM regression paths
}
\examples{
sem_paths(x)
}
