% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_residuals.R
\name{sem_residuals}
\alias{sem_residuals}
\title{Residual correlation matrix}
\usage{
sem_residuals(x, print = TRUE)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of the residual correlation matrix
}
