% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_rsquared.R
\name{sem_rsquared}
\alias{sem_rsquared}
\title{R squared values}
\usage{
sem_rsquared(x, digits = 3, print = TRUE)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{digits}{How many digits to display? (default = 3)}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of R squared values
}
