% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_sig.R
\name{sem_sig}
\alias{sem_sig}
\title{Chi-square significance test}
\usage{
sem_sig(x, print = TRUE)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Model significance tests
}
