% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_tables.R
\name{sem_tables}
\alias{sem_tables}
\title{SEM Tables}
\usage{
sem_tables(
  x,
  standardized = TRUE,
  ci = "standardized",
  ci.level = 0.95,
  digits = 3
)
}
\arguments{
\item{x}{a lavaan model}

\item{standardized}{logical. Include standardized loadings? (default = TRUE)}

\item{ci}{logical. display standardized or unstandardized confidence
intervals? (default = "standardized"). Not needed if standardized=FALSE}

\item{ci.level}{What level of confidence interval to use (default = 0.95)}

\item{digits}{How many digits to display? (default = 3)}
}
\description{
Display all tables relevant for SEM models
}
