% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_factorcor.R
\name{sem_factorcor}
\alias{sem_factorcor}
\title{Latent factor correlations}
\usage{
sem_factorcor(x, standardized = TRUE, factors = c(), print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{standardized}{logical whether to include standardized loadings (default = TRUE)}

\item{factors}{list c() of factors included in the model}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Latent factor correlations
}
\examples{
sem_factorcor(x)
}
