% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_factorloadings.R
\name{sem_factorloadings}
\alias{sem_factorloadings}
\title{Factor loadings}
\usage{
sem_factorloadings(x, standardized = TRUE, print = TRUE)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}

\item{standardized}{logical whether to include standardized loadings (default = TRUE)}

\item{print}{Create a knitr table for displaying as html table (default = TRUE)}
}
\description{
This function will display a table of Factor loadings
}
\examples{
sem_factorloadings(x)
}
