% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_paths.R
\name{sem_paths}
\alias{sem_paths}
\title{SEM regression paths}
\usage{
sem_paths(
  x,
  standardized = TRUE,
  ci = "standardized",
  ci_level = 0.95,
  digits = 3,
  print = TRUE
)
}
\arguments{
\item{x}{a cfa() or sem() lavaan model}

\item{standardized}{logical. Include standardized loadings? (default = TRUE)}

\item{ci}{logical. display standardized or unstandardized confidence
intervals? (default = "standardized"). Not needed if standardized=FALSE}

\item{ci_level}{What level of confidence interval to use (default = 0.95)}

\item{digits}{How many digits to display? (default = 3)}

\item{print}{Create a knitr table for displaying as html table? (default = TRUE)}
}
\description{
This function will display a table of SEM regression paths
}
