% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_tables.R
\name{sem_tables}
\alias{sem_tables}
\title{SEM Tables}
\usage{
sem_tables(
  x,
  standardized = TRUE,
  unstandardized = FALSE,
  ci_level = 0.95,
  digits = 3
)
}
\arguments{
\item{x}{a lavaan model}

\item{standardized}{Logical, indicating whether or not to print standardized
estimates. Standardized estimates are based on "refit" of the model
on standardized data but it will not standardize categorical predictors.
Defualt is TRUE.}

\item{unstandardized}{Logical, indicating whether or not to print
unstandardized estimates. Default is TRUE.}

\item{ci_level}{What level of confidence interval to use (default = 0.95)}

\item{digits}{How many digits to display? (default = 3)}
}
\description{
Display all tables relevant for SEM models. Make sure results='asis' is
specified in the R code chunk setting.
}
